feat: add dashboard
