+++ src/components/Dashboard/index.js
@@ -1,0 +1,3 @@
+import Dashboard from './Dashboard';
+
+export default Dashboard;
