+++ src/components/Form/index.js
@@ -1,0 +1,3 @@
+import Form from './Form';
+
+export default Form;
