+++ src/components/Snackbar/index.js
@@ -1,0 +1,3 @@
+import SnakeBar from './Snackbar';
+
+export default SnakeBar;
