+++ src/components/Transactions/index.js
@@ -1,0 +1,3 @@
+import Transactions from './Transactions';
+
+export default Transactions;
